chromium-ozone-wayland: Upgrade chromium to 142.0.7444.175 version [YOCIMX-9810]

+++ meta-imx-sdk/dynamic-layers/chromium-browser-layer/recipes-browser/chromium/chromium-ozone-wayland_%.bbappend
@@ -7,19 +7,14 @@ SRC_URI:append:imx-nxp-bsp = " \
     file://0004-Fixed-chromium-crash-after-upgrading.patch \
     file://0005-MGS-7765-Blacklist-MSAA-for-GPU-Raster-on-Vivante-GP.patch \
     file://0006-LF-12406-Fixed-webgl-test-fail-for-GL_MAX_SAMPLES-ch.patch \
-    file://0007-Enable-native-GLES2-for-Ozone-wayland.patch \
-    file://0008-Fix-build-fail-after-clang-llvm-upgrade.patch \
-    file://0009-Fix-chromium-crash-when-run-webgl-2.0.0-cts.patch \
-    file://0010-Fix-canvas-test-fail-for-webgl.patch \
-    file://0011-LF-12406-1-Blacklist-disable-program_caching_for_tra.patch \
-    file://0012-Fix-build-error-for-rust-1.90.patch \
-    file://0013-Fix-build-error-for-clang-upgrading.patch \
-    file://0014-Fixed-build-error-caused-by-rust.patch \
-    file://0015-Fix-build-errors-by-rust.patch \
-    file://0016-Disable-GL_ANGLE_instanced_arrays-extension.patch \
-    file://0017-Fix-webgl-regression-after-angle-upgrade.patch \
-    file://0018-MGS-8560-Fixed-webgl-random-failures-with-angle-gles.patch \
-    file://0019-MGS-8643-Fix-chromium-fail-to-run-with-Panthor-drive.patch \
+    file://0007-Fix-build-fail-after-clang-llvm-upgrade.patch \
+    file://0008-Fix-chromium-crash-when-run-webgl-2.0.0-cts.patch \
+    file://0009-Fix-canvas-test-fail-for-webgl.patch \
+    file://0010-LF-12406-1-Blacklist-disable-program_caching_for_tra.patch \
+    file://0011-Disable-GL_ANGLE_instanced_arrays-extension.patch \
+    file://0012-MGS-8560-Fixed-webgl-random-failures-with-angle-gles.patch \
+    file://0013-MGS-8643-Fix-chromium-fail-to-run-with-Panthor-drive.patch \
+    file://fix-SYS_SECCOMP-redefinition.patch \
 "
 VDA_PATCH_SET = " \
     file://0101-V4L2VideoDecoder-Add-macro-use_linux_v4l2.patch \
@@ -38,14 +33,12 @@ VDA_PATCH_SET = " \
     file://0114-V4L2VideoDecoder-support-gpu-import-NV12-format.patch \
     file://0115-VideoDecoderPipeline-Add-resolution-change-support.patch \
     file://0116-V4L2StatefulVideoDecoderBackend-Enlarge-input-buffer.patch \
-    file://0117-V4L2VideoDecoder-Fix-amphion-report-size-mismatch.patch \
-    file://0118-VideoDecoderPipeline-Get-V4L2VideoDecoder-supported-.patch \
-    file://0119-V4L2VideoDecoder-Fix-amphion-cannot-streamoff-after-.patch \
-    file://0120-V4L2VideoDecoder-Set-OUTPUT-format-with-parsed-resol.patch \
-    file://0121-V4L2VideoDecoder-Add-back-HEVC-code-for-stateful-vid.patch \
-    file://0122-V4L2VideoDecoder-Disable-decoder-prioritization-acco.patch \
+    file://0117-V4L2VideoDecoder-Fix-amphion-cannot-streamoff-after-.patch \
+    file://0118-V4L2VideoDecoder-Set-OUTPUT-format-with-parsed-resol.patch \
+    file://0119-V4L2VideoDecoder-Add-back-HEVC-code-for-stateful-vid.patch \
+    file://0120-V4L2VideoDecoder-Disable-decoder-prioritization-acco.patch \
+    file://0121-V4L2VideoDecoder-Fix-cannot-select-V4L2VD.patch \
 "
-SRC_URI:remove = "file://0008-Use-the-correct-path-to-libclang_rt.builtins.a.patch"
 SRC_URI:append:mx8-nxp-bsp = " ${VDA_PATCH_SET}"
 SRC_URI:append:mx95-nxp-bsp = " ${VDA_PATCH_SET}"
 
@@ -97,34 +90,44 @@ CHROMIUM_EXTRA_ARGS:append = " \
     --disable-gpu-rasterization \
     ${CHROMIUM_EXTRA_ARGS_ENABLE_ANGLE} \
 "
-do_copy_clang_library() {
+
+do_copy_clang_library () {
     cp -r "${STAGING_LIBDIR_NATIVE}/clang/latest" "${STAGING_DIR_HOST}${nonarch_libdir}/clang/"
     cd "${STAGING_DIR_HOST}${nonarch_libdir}/clang" || return
 
-    lib_files=$(find . \( -name "libclang_rt.builtins-*" -o -name "liborc_rt-*" \))
+    lib_file=$(find $(find . -maxdepth 1 \! \( -name latest -o -name '.' \)) \( -name "libclang_rt.builtins-*" -o -name "liborc_rt-*" \))
+    echo "lib_file = $lib_file"
     export ARCH="${TARGET_ARCH}"
 
-    target_dir="latest/lib/${ARCH}-unknown-linux-gnu"
-    mkdir -p "$target_dir"
+    mkdir -p "latest/lib/${ARCH}-unknown-linux-gnu"
+    echo "Executing cp $lib_file latest/lib/${ARCH}-unknown-linux-gnu/"
+    cp $lib_file "latest/lib/${ARCH}-unknown-linux-gnu/"
+    cd "latest/lib/${ARCH}-unknown-linux-gnu/" || return
 
-    for lib_file in $lib_files; do
-        # Avoid copying if source and destination are the same
-        dest_path="$target_dir/$(basename "$lib_file")"
-        if [ "$(realpath "$lib_file")" != "$(realpath "$dest_path")" ]; then
-            cp "$lib_file" "$target_dir/"
+    for file in *-"${ARCH}".a *-"${ARCH}hf".a; do
+        if [ -f "$file" ]; then
+            new_name=$(echo "$file" | sed -e "s/-${ARCH}hf//" -e "s/-${ARCH}//")
+            mv "$file" "$new_name"
         fi
-    done
-
-    cd "$target_dir" || return
-
-    for file in *-"${ARCH}".a; do
-        new_name=$(echo "$file" | sed "s/-${ARCH}//")
-        mv "$file" "$new_name"
     done
 
     native_arch_path="${STAGING_LIBDIR_NATIVE}/clang/latest/lib/${ARCH}-unknown-linux-gnu/"
     mkdir -p "$native_arch_path"
+    echo "Executing cp -r $(ls -d *) $native_arch_path"
     cp -r * "$native_arch_path"
+    if [ "${TARGET_ARCH}" != "${BUILD_ARCH}" ]; then
+        cd "${STAGING_LIBDIR_NATIVE}/clang"
+        rm -rf "latest/lib/${BUILD_ARCH}-unknown-linux-gnu"
+        cp -ar latest/lib/linux "latest/lib/${BUILD_ARCH}-unknown-linux-gnu"
+        cd "latest/lib/${BUILD_ARCH}-unknown-linux-gnu"
+
+        for file in *-"${BUILD_ARCH}".a *-"${BUILD_ARCH}hf".a; do
+            if [ -f "$file" ]; then
+                new_name=$(echo "$file" | sed -e "s/-${BUILD_ARCH}hf//" -e "s/-${BUILD_ARCH}//")
+                mv "$file" "$new_name"
+            fi
+        done
+    fi
 }
 
 PACKAGE_ARCH:imx-nxp-bsp = "${MACHINE_SOCARCH}"
